rewrite call function in Auth
“Let’s be clear. What we’re doing here is controversial. We’re adding
this code to make our implementation more testable. We think the
tradeoff is worth it. We are improving the contract. If a user is in
the conn.assigns, we’ll honor it, no matter how it got there. We have
an improved testing story that doesn’t require us to write mocks or any
other elaborate scaffolding.”