Timing out.
We will use Process.send_after to send ourselves a :timedout message.
At first, we will collect messages from the backends as usual using the
timeout of :infinity until we receive the :timedout message. Once we
reach the timeout, we will flip the timeout from :infinity to 0,
indicating that we are no longer willing to wait for the backend
response.